Moved offer-page to component folder
