+++ src/app/components/offer-page/offer-page.component.spec.ts
