+++ src/app/components/offer-page/offer-page.component.ts
